ENG-6592 - Simplify build gradle & some functions (#232)
