--- NeuroID/src/advancedDeviceReactNativeLib/java/com/neuroid/tracker/events/NIDRegisterEventsHelper.kt
@@ -1,64 +1,0 @@
-package com.neuroid.tracker.events
-
-import android.app.Activity
-import android.os.Looper
-import android.view.View
-import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity
-import com.neuroid.tracker.callbacks.*
-import com.neuroid.tracker.service.NIDServiceTracker
-import com.neuroid.tracker.storage.NIDDataStoreManager
-import com.neuroid.tracker.utils.NIDLogWrapper
-import java.util.*
-
-fun registerWindowListeners(activity: Activity) {
-    val viewMainContainer = activity.window.decorView.findViewById<View>(
-        android.R.id.content
-    )
-
-    val callBack = activity.window.callback
-
-    if (callBack !is NIDGlobalEventCallback) {
-        val nidGlobalEventCallback = NIDGlobalEventCallback(
-            callBack,
-            NIDTouchEventManager(viewMainContainer as ViewGroup),
-            viewMainContainer
-        )
-        viewMainContainer.viewTreeObserver.addOnGlobalFocusChangeListener(nidGlobalEventCallback)
-        viewMainContainer.viewTreeObserver.addOnGlobalLayoutListener(
-            nidGlobalEventCallback
-        )
-
-        activity.window.callback = nidGlobalEventCallback
-    }
-}
-
-fun registerTargetFromScreen(
-    activity: Activity,
-    logger: NIDLogWrapper,
-    storeManager: NIDDataStoreManager,
-    registerTarget: Boolean,
-    registerListeners: Boolean,
-    activityOrFragment: String = "",
-    parent: String = "",
-) {
-    val viewMainContainer = activity.window.decorView.findViewById<View>(
-        android.R.id.content
-    ) as ViewGroup
-
-    val hashCodeAct = activity.hashCode()
-    val guid = UUID.nameUUIDFromBytes(hashCodeAct.toString().toByteArray()).toString()
-
-    android.os.Handler(Looper.getMainLooper()).postDelayed({
-        identifyAllViews(
-            viewMainContainer,
-            guid,
-            logger,
-            storeManager,
-            registerTarget,
-            registerListeners,
-            activityOrFragment,
-            parent
-        )
-    }, 300)
-}
