+++ NeuroID/src/reactNative/java/com/neuroid/tracker/events/NIDRegisterEventsHelper.kt
