+++ NeuroID/src/reactNative/java/com/neuroid/tracker/events/NIDTouchEventManager.kt
