+++ app/build.gradle
@@ -14,7 +14,7 @@ if (keystorePropertiesFile.exists()) {
 
 android {
 
-    compileSdkVersion 33
+    compileSdkVersion = 33
 
     packagingOptions {
         pickFirst 'lib/armeabi-v7a/libfbjni.so'
@@ -60,7 +60,8 @@ android {
             keyAlias keystoreProperties['keyAlias']
             keyPassword keystoreProperties['keyPassword']
         }
-    }*/ //Remove comment only when running locally. DO NOT upload changes without commenting these lines again
+    }*/
+    //Remove comment only when running locally. DO NOT upload changes without commenting these lines again
 
     buildTypes {
         debug {
@@ -74,20 +75,21 @@ android {
         }
     }
 
-    flavorDimensions "version"
+    flavorDimensions = ["native", "adv"]
 
     productFlavors {
-        androidLib {
-            dimension "version"
+        android {
+            dimension "native"
         }
-        reactNativeLib {
-            dimension "version"
+        reactNative {
+            dimension "native"
         }
-        advancedDeviceAndroidLib {
-            dimension "version"
+
+        lib {
+            dimension "adv"
         }
-        advancedDeviceReactNativeLib {
-            dimension "version"
+        advancedDeviceLib {
+            dimension "adv"
         }
     }
